Split processor-related stuff into separate class
